Created Services and Controller mapped to the database

+++ src/main/java/com/pancakesunlimited/server/repository/PancakeIngredientRepository.java
@@ -3,5 +3,5 @@ package com.pancakesunlimited.server.repository;
 import com.pancakesunlimited.server.entity.PancakeIngredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PancakeIngredientRepository extends JpaRepository<PancakeIngredient, Long> {
+public interface PancakeIngredientRepository extends JpaRepository<PancakeIngredient, Integer> {
 }
